chore(bridge): move PluginProjectMapper to repositories/mappers/
Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-openagent)
