+++ bridge/app/lib/src/bridge/repositories/mappers/plugin_project_mapper.dart
