+++ bridge/app/lib/src/bridge/routing/open_project_handler.dart
@@ -4,7 +4,7 @@ import "package:sesori_plugin_interface/sesori_plugin_interface.dart";
 import "package:sesori_shared/sesori_shared.dart";
 
 import "../persistence/daos/projects_dao.dart";
-import "plugin_project_mapper.dart";
+import "../repositories/mappers/plugin_project_mapper.dart";
 import "request_handler.dart";
 
 /// Handles `POST /project/open` — opens an existing directory as a project.
