+++ bridge/app/lib/src/bridge/routing/rename_project_handler.dart
@@ -1,7 +1,7 @@
 import "package:sesori_plugin_interface/sesori_plugin_interface.dart";
 import "package:sesori_shared/sesori_shared.dart";
 
-import "plugin_project_mapper.dart";
+import "../repositories/mappers/plugin_project_mapper.dart";
 import "request_handler.dart";
 
 /// Handles `PATCH /project/name` — renames a project.
